Wet mud bricks dry over time
If they are touched by rain or water, they will reset back to stage 0.
If it's got snow on top it won't dry further (but will still reset to 0 if water is adjacent to it)
Water block below do not count since they are not touching the block

Each random tick has a 75% chance of attempting to go to the next drying stage
Check all adjacent mud blocks for their dry-ness stage
If a mud block nearby is in a lower mud stage, we lower the chance the current one will move on to the next drying stage
If a mud block nearby is in a higher drying stage, or is completely dry, the current mud block's chances are bumped up a bit
This is to give the drying a slightly "smoother" look.